いったん、準備でwarnningをけすよ

+++ Engine/Direct3D.cpp
@@ -263,7 +263,7 @@ HRESULT Direct3D::InitShader2D()
 		{ "POSITION", 0, DXGI_FORMAT_R32G32B32_FLOAT, 0, 0,  D3D11_INPUT_PER_VERTEX_DATA, 0 },	//ˆÊ’u
 		{ "TEXCOORD", 0, DXGI_FORMAT_R32G32_FLOAT, 0, sizeof(DirectX::XMVECTOR) , D3D11_INPUT_PER_VERTEX_DATA, 0 },//UVÀ•W
 	};
-	hr = pDevice_->CreateInputLayout(layout.data(), layout.size(),
+	hr = pDevice_->CreateInputLayout(layout.data(), (UINT)layout.size(),
 		pCompileVS->GetBufferPointer(), pCompileVS->GetBufferSize(), &(shaderBundle[SHADER_2D].pVertexLayout_));
 	if (FAILED(hr))
 	{
